Cribbage work
Cards can now be selected for discard, and will be added to the crib.
The computer chooses cards randomly, for now. Added a function to
deck.h/c to find a card by it's index in the deck. Made the cribbage
drawing function into a bunch of smaller specialized functions.