Cards can be moved between tableaus...
It's messy, but it mostly works. The concept is there. It wiggs out when
the tableau is empty, which might be a failure of the deck structure -
it's been a long while since I wrote that and might be worth revisting.
Also, it doesn't draw the "empty" spaces correctly - yet. And of course,
the logic checking isn't there, so cards can be moved around all
willy-nilly. Solid update.