Cleaned up klondike update function
Everything now updates in the update function - which is much cleaner
and nicer to look at then shoving it all in the events function. All the
buttons work - the player can draw cards from the stock, move cards from
the waste to the tableaus, or tableau to tableau. Still wiggs when the
tableaus are empty - I believe I need to revist the linked list
functions I wrote for the deck, particularly seeing what happens when a
card is moved to a NULL deck.