Added in a couple of rules for moving cards...
And I think this might be the logical end of this project - or at least,
this version of it. I don't like the way I built the "deck" data
structure - and can think of so many ways it should have been built
differently. The drawing/keyboard handling elements of this program
are more primative than I would like - Goblin Caves took these ideas and
greatly improved on them. I think what I'll do from here is scrap this
project, and start a new card project using an improved data structure.